requirements.txt
pyqt5
mutagen
requests
pillow